m4/requirements.txt
# M4 CCTV 낙상 감지 시스템 패키지

# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.0.0

# Numerical Computing
numpy>=1.19.0

# Data Processing
pandas>=1.2.0
pyyaml>=5.4.0

# Visualization
matplotlib>=3.3.0

# Progress Bar
tqdm>=4.50.0

# Machine Learning
scikit-learn>=0.24.0

# FastAPI (API 서버용)
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# DB & Environment
supabase>=2.0.0
python-dotenv>=0.19.0

# 추가 유틸리티
requests>=2.25.0
